+++ example/lib/pages/mini_games/multi_scenario/maps/map_biome_2.dart
@@ -29,7 +29,7 @@ class MapBiome2 extends StatelessWidget {
         initDirection: showInEnum.direction,
       ),
       map: WorldMapByTiled(
-        MultiScenarioAssets.mapBiome2,
+        TiledReader.asset(MultiScenarioAssets.mapBiome2),
         forceTileSize: Vector2.all(defaultTileSize),
         objectsBuilder: {
           'sensorLeft': (p) => ExitMapSensor(