+++ example/lib/pages/mini_games/platform/platform_game.dart
@@ -20,7 +20,7 @@ class _PlatformGameState extends State<PlatformGame> {
     return BonfireWidget(
       key: _gameKey,
       map: WorldMapByTiled(
-        'platform/platform_map.tmj',
+        TiledReader.asset('platform/platform_map.tmj'),
         objectsBuilder: {
           'frog': (properties) => FrogEnemy(
                 position: properties.position,