+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # next
+BREAKING GANGES:
 - remove `progress`,`delayToHideProgress`,`progressTransactionBuilder`,`progressTransactionDuration` from `BonfireWidget`.
+- `WorldMapByTiled` now expect `TiledReader` instand of `String`. You can use `TiledReader.asset` or `TiledReader.network`.
 
 # 3.1.2
 - fix bug `AutomaticRandomMovement`. (not stop movement in collision)